snapshot time, then dinner

+++ src/com/eleybourn/bookcatalogue/searches/goodreads/SendOneBookTask.java
@@ -24,16 +24,16 @@ import android.content.Context;
 import android.support.annotation.NonNull;
 
 import com.eleybourn.bookcatalogue.BooksRow;
-import com.eleybourn.bookcatalogue.tasks.BCQueueManager;
 import com.eleybourn.bookcatalogue.BookCatalogueApp;
 import com.eleybourn.bookcatalogue.BookEvents.GrNoIsbnEvent;
 import com.eleybourn.bookcatalogue.BookEvents.GrNoMatchEvent;
-import com.eleybourn.bookcatalogue.cursors.BooksCursor;
+import com.eleybourn.bookcatalogue.database.cursors.BooksCursor;
 import com.eleybourn.bookcatalogue.database.CatalogueDBAdapter;
 import com.eleybourn.bookcatalogue.R;
 import com.eleybourn.bookcatalogue.searches.goodreads.GoodreadsManager.Exceptions.NotAuthorizedException;
 import com.eleybourn.bookcatalogue.searches.goodreads.GoodreadsManager.ExportDisposition;
 import com.eleybourn.bookcatalogue.debug.Logger;
+import com.eleybourn.bookcatalogue.tasks.BCQueueManager;
 import com.eleybourn.bookcatalogue.utils.Utils;
 
 import com.eleybourn.bookcatalogue.taskqueue.QueueManager;
@@ -157,7 +157,7 @@ public class SendOneBookTask extends GenericTask {
 	}
 
 	@Override
-	public long getCategory() {
+	public int getCategory() {
 		return BCQueueManager.CAT_GOODREADS_EXPORT_ONE;
 	}
